requirements.txt
# API 서버 (대시보드용)
fastapi
uvicorn
pydantic

# 수집 및 크롤링
requests
beautifulsoup4
python-dotenv

# 데이터 분석 및 수집기용
pandas
numpy
pymysql
cryptography

# [단타 및 복호화 필수 추가 항목]
websockets
websocket-client
pycryptodome

# 시간 처리
zoneinfo; python_version < "3.9"